feat(workspace-cap): extract sanitization helpers (slice 7 of N preparatory)
ADR-152 section 3 F3.4 preparatory slice. Verbatim port of the prompt-injection
scanning helpers (SYSTEM_PROMPT_FILES, is_system_prompt_file, SANITIZER,
reject_if_injected) from ironclaw workspace mod into dasclaw_workspace_cap.
This removes one direct dependency (crate::safety) from the Workspace type,
unblocking a future full Workspace move.

Closes #704
Refs #674
