+++ crates/dasclaw_workspace_cap/src/lib.rs
@@ -52,6 +52,7 @@ pub mod embeddings;
 pub mod layer;
 pub mod policy;
 pub mod privacy;
+pub mod sanitization;
 pub mod search;
 
 pub mod error;
